update to the main login.py script it now works filly from the cnd, i will make the backup still able to pull from github, but that will be part of an advanced -ish content managment system.
Also added @DeprecationWarning to run_account, this func will be fixed in later verions but it was moved to the main loop as the github pages "cdn" thing i was trying to do was not working well for txt files so i had to hardcode the data.

Lastly, removed get_image as it was the same as get_assets, but it was worse, it did not use headers, and was sprinkeld in the code at random, anyway i have tested with everything moved over to get_assets and it works well, i will prob rename get_assets -> to get_asset as you only get 1 asset at a time.